forsrc 生成的 helloworld 演示项目

+++ src/main/java/com/example/business/user/service/ServiceUpdatePassword.java
@@ -48,7 +48,7 @@ public class ServiceUpdatePassword extends BaseService {
   }
 
   private User getUser(HttpServletRequest request, HttpServletResponse response) {
-    Long userId = getUserId();
+    Integer userId = getUserId();
     return serviceUser.selectByPrimary(request, response, userId);
   }
