+++ src/main/java/com/example/mvc/bean/rep/RepUser.java
@@ -19,8 +19,8 @@ public class RepUser {
   @Data
   public static class UserRow {
 
-    @ApiModelProperty(value = "用户编号", name = "id", dataType = "Long", required = true)
-    private Long id;
+    @ApiModelProperty(value = "用户编号", name = "id", dataType = "Integer", required = true)
+    private Integer id;
 
     @ApiModelProperty(value = "角色类型", name = "roleType", dataType = "Byte", required = true)
     private Byte roleType;
